Support for Android 10
Had to go back to a Service instead of AlarmManager because it didn't work in Android 10.  Now the service will run as a foreground service with a permanant notification.  You can either hide the notification through the usual system UI or keep it open.  You can tap the notification to disable the price updates and re-enable them by tapping on the widget.